+++ modules/apps/commerce/commerce-address-web/src/main/java/com/liferay/commerce/address/web/internal/constants/CommerceCountryScreenNavigationConstants.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.commerce.address.web.internal.servlet.taglib.ui.constants;
+package com.liferay.commerce.address.web.internal.constants;
 
 /**
  * @author Alessio Antonio Rendina
