+++ modules/apps/commerce/commerce-address-web/src/main/resources/META-INF/resources/init.jsp
@@ -25,11 +25,11 @@ taglib uri="http://liferay.com/tld/portlet" prefix="liferay-portlet" %><%@
 taglib uri="http://liferay.com/tld/theme" prefix="liferay-theme" %><%@
 taglib uri="http://liferay.com/tld/ui" prefix="liferay-ui" %>
 
-<%@ page import="com.liferay.commerce.address.web.internal.display.context.CommerceCountriesDisplayContext" %><%@
+<%@ page import="com.liferay.commerce.address.web.internal.constants.CommerceCountryScreenNavigationConstants" %><%@
+page import="com.liferay.commerce.address.web.internal.display.context.CommerceCountriesDisplayContext" %><%@
 page import="com.liferay.commerce.address.web.internal.display.context.CommerceCountriesManagementToolbarDisplayContext" %><%@
 page import="com.liferay.commerce.address.web.internal.display.context.CommerceRegionsDisplayContext" %><%@
 page import="com.liferay.commerce.address.web.internal.display.context.CommerceRegionsManagementToolbarDisplayContext" %><%@
-page import="com.liferay.commerce.address.web.internal.servlet.taglib.ui.constants.CommerceCountryScreenNavigationConstants" %><%@
 page import="com.liferay.commerce.product.model.CommerceChannel" %><%@
 page import="com.liferay.commerce.starter.CommerceRegionsStarter" %><%@
 page import="com.liferay.petra.portlet.url.builder.PortletURLBuilder" %><%@
